Remove files with only duplicated code
On close inspection, "курсовая 20.ipynb" only consists of code already present
in "курсовая 3.0.ipynb". Code duplication is bad, since any potential
improvements also have to be duplicated, which might lead to inconsistency.

Another option could be to remove the duplicated code from "курсовая 3.0.ipynb"
and keep both files, but that doesn't seem to be the intent of it.

Version control already stores the whole history of the project, so there is no
reason to keep old versions of the files. Just keep editing them and making
snapshots regularly.